backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
opencv-python==4.8.1.78
scikit-learn==1.3.2
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
sqlalchemy==2.0.23
alembic==1.12.1
